Struct compiletest_rs::header::TestProps

pub struct TestProps {
    pub error_patterns: Vec<String>,
    pub compile_flags: Vec<String>,
    pub run_flags: Option<String>,
    pub pp_exact: Option<PathBuf>,
    pub aux_builds: Vec<String>,
    pub rustc_env: Vec<(String, String)>,
    pub exec_env: Vec<(String, String)>,
    pub check_lines: Vec<String>,
    pub build_aux_docs: bool,
    pub force_host: bool,
    pub check_stdout: bool,
    pub no_prefer_dynamic: bool,
    pub pretty_expanded: bool,
    pub pretty_mode: String,
    pub pretty_compare_only: bool,
    pub forbid_output: Vec<String>,
    pub revisions: Vec<String>,
    pub incremental_dir: Option<PathBuf>,
    pub must_compile_successfully: bool,
}

Fields
error_patterns: Vec<String> compile_flags: Vec<String> run_flags: Option<String> pp_exact: Option<PathBuf> aux_builds: Vec<String> rustc_env: Vec<(String, String)> exec_env: Vec<(String, String)> check_lines: Vec<String> build_aux_docs: bool force_host: bool check_stdout: bool no_prefer_dynamic: bool pretty_expanded: bool pretty_mode: String pretty_compare_only: bool forbid_output: Vec<String> revisions: Vec<String> incremental_dir: Option<PathBuf> must_compile_successfully: bool
Methods

impl TestProps

fn new() -> Self

fn from_aux_file(&self, testfile: &Path, cfg: Option<&str>) -> Self

fn from_file(testfile: &Path) -> Self

fn load_from(&mut self, testfile: &Path, cfg: Option<&str>)

Load properties from testfile into props. If a property is tied to a particular revision foo (indicated by writing //[foo]), then the property is ignored unless cfg is Some("foo").